解决bug  主页下方导航栏点击“我的” 按钮无法正常显示  多个用户消息未读，读取其中一个用户，未读显示扔存在  缺离线推送功能  写日记时，没添加心情正常保存  拒绝好友申请，离线推送没有提示

+++ index/nanmeishu-index/src/components/index/diary.vue
@@ -190,18 +190,26 @@
         return statu1;
       },
       taleDelete(taleId){
-        console.log(taleId);
-        deleteTale({taleId:taleId},{"accessToken":this.token})
-            .then(res=>{
-              console.log(res);
-              this.$toast.success("删除成功");
-              this.$router.push({
-                path:"/black",
-                query:{
-                  url:"/index"
-                }
+        this.$dialog.confirm({
+          title: '提示',
+          message: '您是否删除此日记？',
+        })
+          .then(() => {
+            deleteTale({taleId:taleId},{"accessToken":this.token})
+              .then(res=>{
+                console.log(res);
+                this.$toast.success("删除成功");
+                this.$router.push({
+                  path:"/black",
+                  query:{
+                    url:"/index"
+                  }
+                })
               })
-            })
+          })
+          .catch(() => {
+
+          });
       }
     },
     created() {
